+++ docstrfmt/docstrfmt.py
@@ -19,6 +19,7 @@ from typing import (
 
 import black
 import docutils
+import docutils.nodes
 from black import Mode
 from blib2to3.pgen2.tokenize import TokenError
 from docutils import nodes
@@ -29,7 +30,6 @@ from docutils.transforms import Transform
 from docutils.utils import Reporter, new_document, unescape
 
 from . import rst_extras
-from .const import SECTION_CHARS
 from .exceptions import InvalidRstError, InvalidRstErrors
 from .rst_extras import _add_directive, generic_role
 from .util import make_enumerator
@@ -151,6 +151,7 @@ class FormatContext:
         self.ordinal_format = "arabic"
         self.section_depth = 0
         self.subsequent_indent = 0
+        self.use_adornments = None
         for key, value in kwargs.items():
             setattr(self, key, value)
 
@@ -243,6 +244,7 @@ class Manager:
         black_config: Mode | None = None,
         docstring_trailing_line: bool = True,
         format_python_code_blocks: bool = True,
+        section_adornments: list[tuple[str, bool]] | None = None,
     ):
         """Initialize the manager."""
         rst_extras.register()
@@ -262,6 +264,7 @@ class Manager:
         self.docstring_trailing_line = docstring_trailing_line
         self.format_python_code_blocks = format_python_code_blocks
         self._in_docstring = False  # for resolving line numbers in code blocks
+        self.section_adornments = section_adornments
 
     def _patch_unknown_directives(self, text: str) -> None:
         """Patch unknown directives and roles into the parser."""
@@ -365,6 +368,46 @@ class Manager:
         )
         return f"{formatted_node}\n"
 
+    def _register_adornments(
+        self, input_lines: list[str], document: nodes.document
+    ) -> None:
+        """Register adornments from source text on all individual sections.
+
+        This method will parse the document tree and original text to-be-formatted, and
+        will register, at the document tree, the current document configuration
+        representing the adornments for parts, chapters and sections on each level of
+        the document. In particular, it will install an attribute called
+        ``adornment-character`` with the character used for underline or overlining the
+        section, and ``adornment-overline``, if the section should be overlined or not.
+
+        input_lines
+            The lines of input (splitted by newline), that we must format.
+
+        document
+            The pre-parsed document tree, that will be modified with new section
+            attributes as described above.
+
+        """
+        for section in document.traverse(nodes.section):
+            title_node = section.next_node(nodes.title)
+            if (
+                title_node
+                and hasattr(title_node, "line")
+                and title_node.line is not None
+            ):
+                underline = input_lines[title_node.line - 1].strip()[0]
+                overline_lineno = title_node.line - 3
+                overline = False
+
+                if overline_lineno >= 0:
+                    candidate_overline = input_lines[overline_lineno].strip()
+                    if candidate_overline and candidate_overline[0] == underline:
+                        overline = True
+
+                # Store this information in the document tree
+                section["adornment-character"] = underline
+                section["adornment-overline"] = overline
+
     def get_code_line(self, code: str, strict: bool = False) -> int:
         """Get the line number of the code in the file."""
         lines = self.original_text.splitlines()
@@ -407,7 +450,9 @@ class Manager:
             "", self.settings.report_level, self.settings.halt_level
         )
         parser.parse(text, doc)
-        self._pre_process(doc, line_offset, len(text.splitlines()))
+        input_lines = text.splitlines()
+        self._pre_process(doc, line_offset, len(input_lines))
+        self._register_adornments(input_lines, doc)
         return doc
 
     def perform_format(
@@ -1643,9 +1688,31 @@ class Formatters:
                 None, chain(self._format_children(node, context)), context, node.line
             )
         )
-        char = SECTION_CHARS[context.section_depth - 1]
-        yield text
-        yield char * len(text)
+        char: str = node.parent["adornment-character"]
+        overline: bool = node.parent["adornment-overline"]
+        if context.manager.section_adornments is not None:
+            try:
+                char, overline = context.manager.section_adornments[
+                    context.section_depth - 1
+                ]
+            except IndexError:
+                context.manager.reporter.error(
+                    f"Section at line {node.line} is at depth "
+                    f"{context.section_depth}, however there are only "
+                    f"{len(context.manager.section_adornments)} adornments to pick "
+                    "from. You must review your inputs or change settings."
+                )
+                raise
+
+        if overline:
+            # section headings with overline are centered
+            yield char * (2 + len(text))
+            yield " " + text
+            yield char * (2 + len(text))
+        else:
+            # sections headings without overline are justified
+            yield text
+            yield char * len(text)
 
     def title_reference(
         self,