+++ CHANGES.rst
@@ -8,7 +8,17 @@ Unreleased
 
 - Added support for Python 3.14.
 - Added ability to skip formatting Python code blocks with the
-  ``--format-python-code-blocks/--no-format-python-code-blocks`` or ``-N`` flag.
+  ``--format-python-code-blocks/--no-format-python-code-blocks`` or ``-N`` command-line
+  argument.
+- Added command-line option ``--section-adornments`` or ``-s`` to control section header
+  adornments.
+- Added command-line argument ``--preserve-adornments`` or ``-pA`` to preserve
+  user-defined section adornments.
+
+**Changed**
+
+- Changed default section heading adornments to match the ones described at
+  https://devguide.python.org/documentation/markup/#sections.
 
 **Fixed**